Поправила структуру проекта
