+++ src/main/java/example/IngredientType.java
@@ -1,0 +1,11 @@
+package example;
+
+/**
+ * Перечисление с типами ингредиентов.
+ * SAUCE – соус
+ * FILLING – начинка
+ */
+public enum IngredientType {
+    SAUCE,
+    FILLING
+}
